Move STACRasterFeatureSerializer to own module
